// web/src/api/generated/model/reportCell.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ReportCell {
  /** @nullable */
  text?: string | null;
  /** @nullable */
  textAr?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  date?: string | null;
}